:truck: Move stops form to new location and rename.
Matching up the the UI panel name.
