+++ src/web/src/modules/travelForm/views/travel-form-create/DetailsTab.vue
@@ -3,7 +3,7 @@
     <PurposeFormCard />
 
     <h3>Details</h3>
-    <StopsFormCreate />
+    <DetailsFormCard />
 
     <h3>Approvals</h3>
     <TravelDetailsForm />
@@ -12,14 +12,14 @@
 
 <script>
 import PurposeFormCard from "@/modules/travelForm/components/create/PurposeFormCard"
-import StopsFormCreate from "@/modules/travelForm/components/StopsFormCreate"
+import DetailsFormCard from "@/modules/travelForm/components/create/DetailsFormCard"
 import TravelDetailsForm from "@/modules/travelForm/components/TravelDetailsForm"
 
 export default {
   name: "DetailsTab",
   components: {
     PurposeFormCard,
-    StopsFormCreate,
+    DetailsFormCard,
     TravelDetailsForm,
   },
 }
